make TextAlign a top level class and use it in both packages
